requirements.txt
# CS4241 RAG — manual stack (no LangChain / LlamaIndex)
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
pypdf>=4.0.0
requests>=2.31.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
torch>=2.1.0
openai>=1.30.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
